do proper conjugation
This makes sure that only the numerator is conjugated
under kpoint symmetry and not the denominators, under the
default setting of --kpoint-symmetry-form conj.

This was a fair bit trickier than expected.  It couldn't be
done by transforming the result of np.einsum because the summation
is over quantities with different denominators.  A second einsum
had to be added for the mirrored Kpoint.